Config updated to use ts
