+++ package.json
@@ -10,11 +10,13 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "^1.11.0",
+    "@skeletonlabs/skeleton": "^1.12.0",
+    "@skeletonlabs/tw-plugin": "^0.1.0",
     "@sveltejs/adapter-auto": "^2.0.0",
     "@sveltejs/adapter-vercel": "^3.0.0",
     "@sveltejs/kit": "^1.5.0",
-    "@tailwindcss/forms": "^0.5.4",
+    "@tailwindcss/forms": "^0.5.6",
+    "@types/node": "^20.5.8",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "postcss-load-config": "^4.0.1",
