+++ postcss.config.cjs
@@ -4,8 +4,7 @@ const autoprefixer = require("autoprefixer");
 const config = {
   plugins: [
     //Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-    tailwindcss(),
-    //But others, like autoprefixer, need to run after,
+    tailwindcss(), //Some plugins, like tailwindcss/nesting, need to run before Tailwind, tailwindcss(), //But others, like autoprefixer, need to run after, //But others, like autoprefixer, need to run after,
     autoprefixer,
   ],
 };
