+++ src/app.postcss
@@ -1,1 +1,7 @@
-html, body { @apply h-full overflow-hidden; }
+@tailwind base;
+html,
+body {
+  @apply h-full overflow-hidden;
+}
+@tailwind components;
+@tailwind utilities;
